requirements.txt
python-telegram-bot[job-queue]==21.7
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.7
apscheduler==3.10.4